# Conflicts:
#	frontend/package.json
#	frontend/webApp/vite.config.ts

+++ frontend/shared/interop-sample/domain/src/commonMain/kotlin/interopSample/models/Weather.kt
@@ -1,7 +1,10 @@
 package interopSample.models
 
+import utils.BigDecimal
+import kotlin.js.JsExport
+
 @JsExport
 data class Weather(
-    val temperature: Double,
+    val temperature: BigDecimal,
     val temperatureUnit: String
 )
